backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1; python_version < "3.9"
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
dj-rest-auth==4.0.0
Django==4.2.2
django-allauth==0.52.0
django-cors-headers==4.1.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.2
idna==3.6
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
gunicorn==20.1.0
whitenoise==6.4.0
psycopg2-binary==2.9.9
django-sslserver==0.22
